+++ src/ebn/ebn.jl
@@ -1,22 +1,4 @@
 abstract type AbstractNetwork end
 
-# include("cpd/conditionalprobabilitydistribution.jl")
-# include("enhancedbn.jl")
-# include("bayesnet.jl")
-# include("credalnet.jl")
-
-# include("discretization/discretize_node.jl")
-# include("discretization/discretize_net.jl")
-
-# include("reduction.jl")
-
-# include("trasmission/trasmission_node.jl")
-# include("trasmission/trasmission_net.jl")
-
-# include("evaluation/evaluate_node.jl")
-# include("evaluation/evaluate_net.jl")
-
-
-# include("new_enhancedbn.jl")
 include("../util/node_verification.jl")
-include("new_enhancedbn.jl")
+include("enhancedbn.jl")
