+++ src/nodes/nodes.jl
@@ -17,8 +17,6 @@ const AbstractContinuousInput = Union{UnivariateDistribution,Tuple{Real,Real},Un
 
 const AbstractDiscreteProbability = Union{<:Real,AbstractVector{<:Real}}
 
-include("../util/wrap.jl")
-
 abstract type AbstractDiscretization end
 
 """ ExactDiscretization
@@ -82,55 +80,6 @@ function _extreme_points_states_probabilities(states::Dict{Symbol,<:AbstractDisc
     return map(val -> Dict(keys(states) .=> val), v.points.points)
 end
 
-# function _get_position(nodes::AbstractVector{<:AbstractNode})
-#     adj_matrix = get_adj_matrix(nodes)
-#     pos = spring(adj_matrix; iterations=1000)
-#     return pos
-# end
-
-# function get_adj_matrix(nodes::AbstractVector{<:AbstractNode})
-#     ordered_list = _order_node(nodes)
-#     n = length(ordered_list)
-#     adj_matrix = zeros(n, n)
-#     for i in range(1, n)
-#         for j in range(1, n)
-#             if !isa(ordered_list[j], RootNode) && ordered_list[i] ∈ ordered_list[j].parents
-#                 adj_matrix[i, j] = 1
-#             end
-#         end
-#     end
-#     return sparse(adj_matrix)
-# end
-
-# function _get_edges(adj_matrix::SparseMatrixCSC)
-#     n = size(adj_matrix)
-#     edge_list = Vector{Tuple{Int64,Int64}}()
-#     for i in range(1, n[1])
-#         for j in range(1, n[2])
-#             if adj_matrix[i, j] != 0
-#                 push!(edge_list, (i, j))
-#             end
-#         end
-#     end
-#     return edge_list
-# end
-
-# function _order_node(nodes::AbstractVector{<:AbstractNode})
-#     root = filter(n -> isa(n, RootNode), nodes)
-#     list = setdiff(nodes, root)
-#     while !isempty(list)
-#         new_root = filter(x -> all(x.parents .∈ [root]), list)
-#         root = append!(root, new_root)
-#         list = setdiff(list, new_root)
-#     end
-#     return root
-# end
-
-# include("../util/node_verification.jl")
-# include("root.jl")
-# include("child.jl")
-# include("functional.jl")
-# include("common.jl")
 include("../util/probabilities_verification.jl")
 include("root.jl")
 include("child.jl")
